updating as per feedback
